+++ CHANGELOG.md
@@ -1,0 +1,8 @@
+=======
+History
+=======
+
+0.1.0 (2016-07-07)
+------------------
+
+* First release on PyPI.
