+++ MANIFEST.in
@@ -1,0 +1,8 @@
+include CHANGELOG.md
+include LICENSE.md
+include README.md
+
+recursive-exclude * __pycache__
+recursive-exclude * *.py[co]
+
+recursive-include *.md *.jpg *.png *.gif
