+++ captain_comeback/__init__.py
@@ -1,0 +1,5 @@
+# coding: utf-8
+
+__author__ = 'Thomas Orozco'
+__email__ = 'thomas@aptible.com'
+__version__ = '0.1.0'
