+++ captain_comeback/restart/messages.py
@@ -1,0 +1,9 @@
+# coding:utf-8
+class RestartRequestedMessage(object):
+    def __init__(self, cg):
+        self.cg = cg
+
+
+class RestartCompleteMessage(object):
+    def __init__(self, cg):
+        self.cg = cg
