+++ tox.ini
@@ -1,0 +1,18 @@
+[tox]
+envlist = py27, py34, py35, flake8
+
+[testenv:flake8]
+basepython=python
+deps=flake8
+commands=flake8 captain_comeback
+
+[testenv]
+setenv =
+    PYTHONPATH = {toxinidir}:{toxinidir}/captain_comeback
+
+commands = python setup.py test
+
+; If you want to make tox run the tests with the same versions, create a
+; requirements.txt with the pinned versions and uncomment the following lines:
+; deps =
+;     -r{toxinidir}/requirements.txt
